OMS-N: integrate OverlayManagerService into framework [6/14]
Hand over ownership of overlays to OverlayManagerService.

Changes to a package's overlays are propagated using the activity life
cycle. Affected activities will be recreated as needed. This provides a
well-defined point to modify an application's assets while the
application is paused.

Consolidate how overlays targeting the system and overlays targeting
regular applications are handled. Previously, system overlays were
handled as a special case. Now, everything is handled identically. As a
side effect, the call to idmap --scan during Zygote boot has become
obsolete and is removed.

Deprecate and remove use of SCAN_TRUSTED_OVERLAY from
PackageManagerService. Previously, the flag was used to restrict what
paths overlays were allowed to be installed in. Now, overlay packages
are first class packages and it is up to the OverlayManagerService to
decide whether to use the overlay or not.

Deprecate ApplicationInfo.resourcesDirs. The old constant value is
obsolete now that the set of overlays can change dynamically. Query the
OverlayManagerService instead.

Bug: 31052947

Co-authored-by: Martin Wallgren <martin.wallgren@sonymobile.com>
Signed-off-by: Zoran Jovanovic <zoran.jovanovic@sonymobile.com>
AOSP Change-Id: Idc96dae6fc075d5373aa055bbf50e919136d7353

Conflicts:
	core/java/android/content/pm/PackageParser.java
	services/core/java/com/android/server/pm/PackageManagerService.java

Change-Id: I175c0af62a59aed9ac94147683e90c5a77cd40d9
Signed-off-by: theh2o64@gmail.com <theh2o64@gmail.com>